skill_matrix_api/requirements.txt
django-ninja
ipdb